Revert "Started single_pe stream"
